+++ src/sRef.c
@@ -10100,14 +10100,14 @@ bool sRef_isFixedArray (sRef p_s) /*@*/ {
   return ( ctype_isFixedArray (c) );
 }
 
-long int sRef_getArraySize (sRef p_s) /*@*/ {
+size_t sRef_getArraySize (sRef p_s) /*@*/ 
+{
   ctype c;
   llassert (sRef_isFixedArray(p_s) );
   DPRINTF (( message ("sRef_getArraySize getting array size for %s", sRef_unparse(p_s) )  ));
   
   c = sRef_getType (p_s);
-
-  return (ctype_getArraySize (c) );
+  return (ctype_getArraySize (c));
 }
 
 void sRef_setValue (sRef s, multiVal val)
